# Improve Performance of Plug.Conn.Cookies.decode/1