# Description
This pull request optimizes the Plug.Conn.Cookies.decode/1 function to improve its performance.

This PR addresses issue [#1227](https://github.com/elixir-plug/plug/issues/1227).