@josevalim 2024-06-20
Hi @isavita! I have already pushed a fix. I compared it with yours and the current version in main is faster. Please double check you see the same result.